Universal Remote/quick-commerce-backend/requirements.txt
Flask
google-cloud-dialogflow-cx
google-cloud-firestore
google-cloud-secret-manager
google-cloud-pubsub
requests
gunicorn